+++ src/main/java/chess/Application.java
@@ -1,10 +1,8 @@
 package chess;
 
-import chess.controller.Controller;
-
 public class Application {
     public static void main(String[] args) {
-        Controller controller = new Controller();
+        ChessGame controller = new ChessGame();
         controller.run();
     }
 }
